+++ src/DefaultCodes.ts
@@ -1,6 +1,3 @@
 import ErrorCode from "./ErrorCode";
 
-export const CODE_MISC_INTERNAL_ERROR = new ErrorCode("misc/internal-error", `An internal error was caught but the operation couldn't be completed`, {
-  severity   : "error",
-  displayName: "InternalError",
-});
+export const CODE_MISC_INTERNAL_ERROR = new ErrorCode("misc/internal-error", `An internal error was caught but the operation couldn't be completed`);