+++ src/ErrorCode.ts
@@ -1,15 +1,19 @@
-import SendableError, {isSendableError} from "./SendableError";
-import {ErrorOptions, ScopedValue}      from "./Types";
+import SendableError, {isSendableError, SendableErrorDetails} from "./SendableError";
+//import {ErrorOptions}                   from "./Types";
 import {CODE_MISC_INTERNAL_ERROR}       from "./DefaultCodes";
 
-export default class ErrorCode {
+export interface ErrorCodeOptions {
+
+}
+
+export default class ErrorCode<D extends SendableErrorDetails = {}> {
 
   readonly prefix: string;
   readonly id: string;
-  readonly defaultMessage: ScopedValue<string>;
-  readonly options: ErrorOptions;
+  readonly defaultMessage: string;
+  //readonly options: ErrorOptions;
 
-  constructor(id: string, defaultMessage?: ScopedValue<string>, options?: ErrorOptions) {
+  constructor(id: string, defaultMessage?: string, /*options?: ErrorOptions*/) {
     const parts = id.split("/");
     if (parts.length !== 2) {
       console.error(`[sendable-error] Invalid error code '${id}' provided, must be of form 'prefix/descriptive-name'`);
@@ -17,25 +21,29 @@ export default class ErrorCode {
     this.prefix         = id.split("/")[0];
     this.id             = id;
     this.defaultMessage = defaultMessage || "An unknown error occurred";
-    this.options        = options || {};
+    //this.options        = options || {};
   }
 
   /**
    * Does this error have this code?
    */
-  is(error: Error | SendableError): boolean {
-    return Boolean(error && "code" in error && error.getCode().id === this.id);
+  is(error: any): error is SendableError {
+    if (isSendableError(error)) {
+      return error.getCode().id === this.id;
+    }
+
+    return false;
   }
 
   /**
    * Extends an existing error code
    */
-  extend(id: string, defaultMessage = undefined, options = {}) {
-    const mixedOptions = {
+  private extend(id: string, defaultMessage = undefined, options = {}) {
+    /*const mixedOptions = {
       ...this.options,
       ...options,
-    };
-    const code         = new ErrorCode(id, defaultMessage || this.defaultMessage, mixedOptions);
+    };*/
+    const code         = new ErrorCode(id, defaultMessage || this.defaultMessage, /*mixedOptions*/);
     if (code.prefix !== this.prefix) {
       console.error("[sendable-error] Extension doesn't share same prefix, this doesn't seem right", {
         originalPrefix: this.prefix,