+++ src/Logger.ts
@@ -1,10 +1,9 @@
-import {ErrorOptions} from "./Types";
 import SendableError  from "./SendableError";
 
 export type ErrorLogger = (options: ErrorLoggerOptions) => void;
 
 export interface ErrorLoggerOptions {
-  options: Required<ErrorOptions>;
+  //options: Required<ErrorOptions>;
   source: string;
   message: string;
   error: SendableError;
@@ -13,13 +12,8 @@ export interface ErrorLoggerOptions {
   providedInfo: any;
 }
 
-export const defaultErrorLogger: ErrorLogger = ({options, source, message, error, info}) => {
-  if (options.severity !== "error") {
-    console[options.severity](source || `ErrorUtils::log`, message, info);
-  } else {
-    console.error(source || `ErrorUtils::log`, message, info);
-    console.error(error);
-  }
+export const defaultErrorLogger: ErrorLogger = ({source, message, error, info}) => {
+    console.error(source || `SendableError`, message, info, error.stack);
 };
 
 let currentLogger = defaultErrorLogger;