+++ src/index.ts
@@ -1,15 +1,16 @@
-export * from "./Types";
+//export * from "./Types";
 export * from "./DefaultCodes";
 export * from "./Logger";
-export * from "./Parser";
+//export * from "./Parser";
 
-export {default as ErrorCode} from "./ErrorCode";
+export {default as ErrorCode, ErrorCodeOptions} from "./ErrorCode";
 
-import {default as SendableError, getTraceId} from "./SendableError";
+import {default as SendableError, getTraceId, isSendableError} from "./SendableError";
 
 export default SendableError;
 
 export {
   SendableError,
   getTraceId,
-}
+  isSendableError
+};