+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 0.1.0
+
+## Breaking Changes
+
+- strip API back to basics -- complex features will come back over time
+- remove visibility/scoped value API for now
+- `messages` -> `message`
+- remove builder properties from `SendableError`. If you want to edit an error use `SendableError.of(oldError).message("...")`
+
 # 0.0.x
 
 ## Features